requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.11.0
fonttools==4.29.1
frozenlist==1.3.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
multidict==5.2.0
nextcord==2.0.0a6
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
pygame==2.1.2
pyparsing==3.0.7
pypresence==4.2.1
python-dateutil==2.8.2
python-dotenv==0.19.2
PyTouchBar==0.2.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
yarl==1.7.2